refactor: moving runserver to the server package
